requirements.txt
python-binance>=1.0.17
ta>=0.10.2
pandas>=1.5.3
python-dotenv>=1.0.1
flask
